# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Photography Booking Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin profile
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
    put:
      operationId: updateMe
      tags: [auth]
      summary: Update admin profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List all booking requests
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: eventType
          in: query
          schema:
            type: ["string", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/assign:
    post:
      operationId: assignPhotographer
      tags: [bookings]
      summary: Assign a photographer to a booking day
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignPhotographerBody"
      responses:
        "200":
          description: Assignment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /inquiries:
    get:
      operationId: getInquiries
      tags: [inquiries]
      summary: List all inquiry requests
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Create a new inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryBody"
      responses:
        "201":
          description: Created inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /inquiries/{id}:
    get:
      operationId: getInquiry
      tags: [inquiries]
      summary: Get inquiry details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inquiry details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
...
  /packages:
    get:
      operationId: getPackages
      tags: [packages]
      summary: List all packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    put:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageBody"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /addons:
    get:
      operationId: getAddons
      tags: [packages]
      summary: List all add-ons
      responses:
        "200":
          description: List of add-ons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Addon"
    post:
      operationId: createAddon
      tags: [packages]
      summary: Create an add-on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddonBody"
      responses:
        "201":
          description: Created add-on
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Addon"
  /addons/{id}:
    put:
      operationId: updateAddon
      tags: [packages]
      summary: Update an add-on
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddonBody"
      responses:
        "200":
          description: Updated add-on
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Addon"
    delete:
      operationId: deleteAddon
      tags: [packages]
      summary: Delete an add-on
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"
      required:
        - token
        - admin
    Admin:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - name
        - email
        - role
    UpdateAdminBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
        status:
          type: string
          enum: [new, pending, assigned, confirmed]
        assignedPhotographerId:
          type: ["integer", "null"]
        assignedPhotographerName:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        packageName:
          type: ["string", "null"]
        totalPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clientName
        - clientPhone
        - clientEmail
        - eventType
        - eventDate
        - location
        - status
        - createdAt
    BookingDay:
      type: object
      properties:
        dayNumber:
          type: integer
        date:
          type: string
        details:
          type: string
        photographerId:
          type: ["integer", "null"]
        photographerName:
          type: ["string", "null"]
      required:
        - dayNumber
        - date
        - details
    BookingDetail:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
        status:
          type: string
          enum: [new, pending, assigned, confirmed]
        assignedPhotographerId:
          type: ["integer", "null"]
        assignedPhotographerName:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        packageName:
          type: ["string", "null"]
        addons:
          type: array
          items:
            type: string
        totalPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        days:
          type: array
          items:
            $ref: "#/components/schemas/BookingDay"
        createdAt:
          type: string
      required:
        - id
        - clientName
        - clientPhone
        - clientEmail
        - eventType
        - eventDate
        - location
        - status
        - days
        - addons
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
        packageId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - clientName
        - clientPhone
        - clientEmail
        - eventType
        - eventDate
        - location
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        assignedPhotographerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    AssignPhotographerBody:
      type: object
      properties:
        dayNumber:
          type: integer
        photographerId:
          type: integer
      required:
        - dayNumber
        - photographerId
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - message
        - createdAt
    CreateInquiryBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        message:
          type: string
      required:
        - name
        - phone
        - message
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        duration:
          type: ["string", "null"]
      required:
        - id
        - name
        - category
        - price
    CreatePackageBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        duration:
          type: ["string", "null"]
      required:
        - name
        - category
        - price
    UpdatePackageBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        duration:
          type: ["string", "null"]
    Addon:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        description:
          type: ["string", "null"]
      required:
        - id
        - name
        - price
    CreateAddonBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        description:
          type: ["string", "null"]
      required:
        - name
        - price
    UpdateAddonBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        description:
          type: ["string", "null"]