# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Booking Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
}