# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Booking Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  clientName: string;
  clientPhone: string;
  clientEmail: string;
  eventType: string;
  eventDate: string;
  /** @nullable */
  endDate?: string | null;
  location: string;
  status: BookingStatus;
  /** @nullable */
  assignedPhotographerId?: number | null;
  /** @nullable */
  assignedPhotographerName?: string | null;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  packageName?: string | null;
  /** @nullable */
  totalPrice?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}