# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Booking Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  clientName: string;
  clientPhone: string;
  clientEmail: string;
  eventType: string;
  eventDate: string;
  /** @nullable */
  endDate?: string | null;
  location: string;
  /** @nullable */
  packageId?: number | null;
  /** @nullable */
  notes?: string | null;
}