# lib/api-zod/src/generated/types/createPackageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Booking Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePackageBody {
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  duration?: string | null;
}