# lib/api-zod/src/generated/types/updatePackageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photography Booking Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePackageBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  duration?: string | null;
}